Tue, Feb 1, 2022, 12:01 PM +00:00

+++ filter-less/helpers.c
@@ -10,7 +10,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     {
         for (int j = 0; j < width; j++)
         {
-            double average = (image[i][j].rgbtRed + image[i][j].rgbtGreen + image[i][j].rgbtBlue)/3;
+            int average = (image[i][j].rgbtRed + image[i][j].rgbtGreen + image[i][j].rgbtBlue)/3;
             average = round(average);
 
             //set the channel values to the average
